pyproject.toml
[tool.poetry]
name = "web-api"
authors = ["pavel-97"]
version = "1"
description = "test"

[tool.poetry.dependencies]
python = "3.12.10"
uvicorn = "0.32"
fastapi = "0.115"
sqlalchemy = "2.0.36"
alembic = "1.14.0"
asyncpg = "0.30.0"
psycopg2-binary = "2.9.10"
httpx = "0.28"
pydantic-settings = "2.6.1"
pytest = "8.3.4"
pytest-asyncio = "0.24.0"
python-multipart = "0.0.19"
ruff = "0.8.2"

[tool.ruff]
line-length = 79

[tool.pytest.ini_options]
pythonpath = [
    ".", "src"
]
asyncio_mode = "auto"